- Significant work on commands and spells - Re-arranged the Loaders into sub-folders - Moved some static data into Attributes and updated relevant enums - Implemented Check functions and converted tons of if-sendtochar blocks to use them

+++ trunk/SmaugCS/Extensions/SavingThrowDataExtensions.cs
@@ -18,7 +18,7 @@ namespace SmaugCS
 
         public static bool CheckSaveVsWandRod(this SavingThrowData savingThrow, int level, CharacterInstance victim)
         {
-            if (victim.Immunity.IsSet((int)ResistanceTypes.Magic))
+            if (victim.Immunity.IsSet(ResistanceTypes.Magic))
                 return true;
 
             int save = 50 + (victim.Level - level - victim.SavingThrows.SaveVsWandRod) * 5;
@@ -42,7 +42,7 @@ namespace SmaugCS
 
         public static bool CheckSaveVsSpellStaff(this SavingThrowData savingThrow, int level, CharacterInstance victim)
         {
-            if (victim.Immunity.IsSet((int)ResistanceTypes.Magic))
+            if (victim.Immunity.IsSet(ResistanceTypes.Magic))
                 return true;
             if (victim.IsNpc() && level > 10)
                 level -= 5;
